Split up main Web View into pieces
