+++ src/__tests__/components/ScriptureNavControls.test.tsx
@@ -1,0 +1,56 @@
+/** @file Unit tests for components/ScriptureNavControls.tsx. */
+/// <reference types="jest" />
+/// <reference types="@testing-library/jest-dom" />
+
+import type { SerializedVerseRef } from '@sillsdev/scripture';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import { useLocalizedStrings, useRecentScriptureRefs } from '@papi/frontend/react';
+import ScriptureNavControls from '../../components/ScriptureNavControls';
+
+const defaultScrRef: SerializedVerseRef = { book: 'GEN', chapterNum: 1, verseNum: 1 };
+
+describe('ScriptureNavControls', () => {
+  beforeEach(() => {
+    jest.mocked(useLocalizedStrings).mockReturnValue([{}, false]);
+    jest.mocked(useRecentScriptureRefs).mockReturnValue({
+      recentScriptureRefs: [],
+      addRecentScriptureRef: jest.fn(),
+    });
+  });
+
+  it('shows the book chapter control', () => {
+    render(
+      <ScriptureNavControls
+        scrRef={defaultScrRef}
+        handleSubmit={() => {}}
+        scrollGroupId={undefined}
+        onChangeScrollGroupId={() => {}}
+      />,
+    );
+
+    expect(screen.getByTestId('book-chapter-control')).toBeInTheDocument();
+  });
+
+  it('calls handleSubmit and addRecentScriptureRef when the verse picker submits', async () => {
+    const mockHandleSubmit = jest.fn();
+    const mockAddRecentRef = jest.fn();
+    jest.mocked(useRecentScriptureRefs).mockReturnValue({
+      recentScriptureRefs: [],
+      addRecentScriptureRef: mockAddRecentRef,
+    });
+    render(
+      <ScriptureNavControls
+        scrRef={defaultScrRef}
+        handleSubmit={mockHandleSubmit}
+        scrollGroupId={undefined}
+        onChangeScrollGroupId={() => {}}
+      />,
+    );
+
+    await userEvent.click(screen.getByRole('button', { name: /submit reference/i }));
+
+    expect(mockHandleSubmit).toHaveBeenCalledWith(defaultScrRef);
+    expect(mockAddRecentRef).toHaveBeenCalledWith(defaultScrRef);
+  });
+});
